CPS-65 changed /summary to just /communities
it's more REST, you get a list of communities (and also a bunch of other stuff that probably belongs to /config or something)
